+++ source/IGrid.hpp
@@ -17,7 +17,7 @@ class IGrid
         virtual int movePhotonAtRandomDepth(Photon& ph, Random *ran) const = 0;
         virtual void peeloff(Photon ph, Observer& observer, DustCRef dust) const = 0;
         virtual double computeMatterMass() const = 0;
-        virtual std::uint32_t cellId(const Vector3d& position) const = 0;
+        virtual std::uint64_t cellId(const Vector3d& position) const = 0;
 };
 
 #endif
