+++ source/Photon.hpp
@@ -11,13 +11,13 @@
 class Photon
 {
     public:
-        Photon(Vector3d const &position, std::uint32_t cellId, Direction3d const &dir, double weight, std::uint32_t nscat, double fi=1.0, double fq=0.0, double fu=0.0, double fv=0.0);
+        Photon(Vector3d const &position, std::uint64_t cellId, Direction3d const &dir, double weight, std::uint32_t nscat, double fi=1.0, double fq=0.0, double fu=0.0, double fv=0.0);
         // photon scattering
         double Scatt(DustCRef dust, Direction3d const & dir, Random* ran);
         void Scatt( Model const &m, Directions const &dirs, IGridCRef grid, std::vector<Observer>& observers, Random* ran);
         void Stokes(DustCRef dust, Direction3d const &dir, double calpha, bool fDir, Random* ran);
 
-        void Move(double t, std::uint32_t cellId)
+        void Move(double t, std::uint64_t cellId)
         {
             cellId_ = cellId;
             pos_ = pos_ + t * dir_.vector();
@@ -46,9 +46,9 @@ class Photon
         {	return weight_;	}
         double weight() const
         {	return weight_;	}
-        std::uint32_t& cellId()
+        std::uint64_t& cellId()
         {	return cellId_;	}
-        std::uint32_t cellId() const
+        std::uint64_t cellId() const
         {	return cellId_;	}
 
         bool termination() const
@@ -58,7 +58,7 @@ class Photon
         Vector3d	pos_;	// outpoint
         Direction3d	dir_;	// vector of the direction
         std::uint32_t 	nscat_;
-        std::uint32_t    cellId_;
+        std::uint64_t    cellId_;
         double weight_;
         double fi_, fq_, fu_, fv_;
 };
